refactor: overhaul avatar component structure and props management
- Removed legacy avatar controller and shadow components.
- Introduced new functional avatar component with improved props handling.
- Implemented a new controller for avatar to manage state and props validation.
- Added color, initials, label, and src props with type safety.
- Updated styles for avatar component to enhance layout and appearance.
- Created a new skeleton component with integrated click button functionality.
- Established a robust props validation system for various component properties.
- Enhanced documentation for props and their usage patterns.
