--- packages/components/src/components/_skeleton/web-components/avatar/style.scss
@@ -1,26 +1,0 @@
-@use '../../../@shared/mixins' as *;
-@use '../../../../styles/global' as *;
-@use '../../../host-display-block' as *;
-
-@layer kol-component {
-	.kol-avatar,
-	.kol-avatar__image,
-	.kol-avatar__initials {
-		border-radius: 50%;
-	}
-
-	.kol-avatar {
-		/* Visible with forced colors  */
-		outline: transparent solid to-rem(1);
-		width: to-rem(100);
-		height: to-rem(100);
-
-		display: flex;
-
-		&__initials {
-			margin: auto;
-			background-color: inherit;
-			font-size: to-rem(40);
-		}
-	}
-}
