--- packages/components/src/components/avatar/controller.ts
@@ -1,20 +1,0 @@
-const formatNameAsInitial = (name: string): string => {
-	if (name.length === 0) {
-		return '';
-	}
-
-	return name[0].toUpperCase();
-};
-
-export const formatLabelAsInitials = (label: string): string => {
-	const names = label.split(/\s+/); // split by any whitespace characters
-	const first = names[0];
-	const last = names[names.length - 1];
-
-	// names might consist of only one word
-	if (names.length >= 2 && first && last) {
-		return `${formatNameAsInitial(first)}${formatNameAsInitial(last)}`;
-	}
-
-	return formatNameAsInitial(label);
-};
