+++ packages/components/src/internal/functional-components/skeleton/controller.ts
@@ -1,6 +1,6 @@
-import { Log } from '../../../../../schema';
-import type { CountProp, LabelProp, NameProp } from '../../schema/props';
-import { countProp, labelProp, nameProp, withValidPropValue } from '../../schema/props';
+import { Log } from '../../../schema';
+import type { CountProp, LabelProp, NameProp } from '../../props';
+import { countProp, labelProp, nameProp, withValidPropValue } from '../../props';
 import { BaseController } from '../base-controller';
 import { ClickButtonController } from '../click-button/controller';
 import type { ControllerInterface, ResolvedInputProps } from '../generic-types';
