+++ packages/components/src/internal/props/count.ts
