+++ packages/components/src/internal/props/initials.ts
