+++ packages/components/src/internal/props/label.ts
