+++ packages/components/src/internal/props/show.ts
